rename: src/app/components -> src/components
폴더 위치 변경
