+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import MemoList from './components/common/MemoList';
+import MemoList from '@/components/memo/MemoList';
 import { MEMOS } from '@/data/mocks';
 
 export default function Home() {
