+++ src/components/layout/sidebar/CategoryItem.tsx
