+++ src/components/memo/MemoCard.tsx
